\name{allele.marq}

\alias{allele.marq}

\title{alleles at markers}

\description{
the function finds the alleles present at each marker given a set of haplotypes.
It can be viewed as an internal function. 
The user does not have to call it by himself.
}

\usage{
allele.marq(hap)
}

\arguments{
  \item{hap}{character matrix (n*p) of haplotypes of individuals.}
}

\value{
The value returned is a list of alleles for each marker.
}

\references{publication to be submitted:C.Cierco-Ayrolles,S.Dejean,A.Legarra,H.Gilbert,
T.Druet,F.Ytournel,D.Estivals,N.Oumouhou and B.Mangin,
Combining linkage analysis and linkage disequilibrium for QTL fine mapping in animal pedigrees.}

\author{S. Dejean, N. Oumouhou, D. Estivals, B. Mangin }

\keyword{models}








