\name{haldanem1}

\alias{haldanem1}

\title{recombination rate}

\description{
The function calculates recombination rates using Haldane distance.
It can be viewed as an internal function. 
The user does not have to call it by himself.
}

\usage{
haldanem1(distance)
}

\arguments{
  \item{distance}{vector of distances in Morgan.}
}

\value{
the returned value  is a vector of recombination rates.
}

\references{
Haldane, J.B.S. {\it J. Genet.} (1919), 8:299-309.

publication to be submitted: C. Cierco-Ayrolles, S. Dejean, A. Legarra, H. Gilbert,
T. Druet, F. Ytournel, D. Estivals, N. Oumouhou and B. Mangin.
Combining linkage analysis and linkage disequilibrium for QTL fine mapping in animal pedigrees.}


\author{S. Dejean, N. Oumouhou, D. Estivals, B. Mangin }

\keyword{models}
