\name{HAPim-package}

\alias{HAPim-package}

\alias{HAPim}

\docType{package}

\title{HAPim}

\description{
The package provides a set of functions whose aim is to propose 4 methods of QTL detection:

-HAPim LD is an interval-mapping method designed for unrelated individuals 
with no family information and is based on a maximum-likelihood calculation.

-HAPim LDL is an interval-mapping method which uses family information of individuals and 
combines linkage and linkage disequilibrium.It is based on a maximum-likelihood calculation.

-Haplomax gives an analysis of variance with a sire effect and a haplotype effect. 

-Haplomax HS gives an analysis of variance with a sire effect and a haplotype effect in half-sire design.

Fundings for the package development were provided to the LDLmapQTL project by the ANR GENANIMAL program and APIS-GENE.

}


\details{
\tabular{ll}{
Package: \tab HAPim\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2008-05-28\cr
License: \tab none\cr
}
}

\author{S. Dejean, N. Oumouhou, D. Estivals, B. Mangin 

 Maintainer: Brigitte Mangin <brigitte.mangin@toulouse.inra.fr> }

\references{publication}


\keyword{models}



