\name{pi.hap}

\alias{pi.hap}

\title{pi.hap}

\description{
The function can be viewed as an internal function. 
The user does not have to call it by himself.
}

\usage{
pi.hap(freq.marq, res.structure)
}


\arguments{
  \item{freq.marq}{results provided by \code{freq.all()} function, list of numeric objects.}
  \item{res.structure}{results provided by \code{structure()} function, list of numeric objects.}
}


\value{
The value returned is a list of numeric vectors.
}

\references{publication to be submitted:C.Cierco-Ayrolles,S.Dejean,A.Legarra,H.Gilbert,
T.Druet,F.Ytournel,D.Estivals,N.Oumouhou and B.Mangin,
Combining linkage analysis and linkage disequilibrium for QTL fine mapping in animal pedigrees.}

\author{S. Dejean, N. Oumouhou, D. Estivals, B. Mangin }

\seealso{\code{\link{freq.all}},{\code{\link{structure}}} }

\keyword{models}
