\name{structure}

\alias{structure}

\title{Structure}

\description{
The function can be viewed as an internal function. 
The user does not have to call it by himself.
}

\usage{
structure(nbre.marq, nbre.all.marq)
}

\arguments{
  \item{nbre.marq}{number of markers.}
  \item{nbre.all.marq}{number of alleles at markers.}
}


\value{
The value returned is a list of objects.
}

\references{publication to be submitted:C.Cierco-Ayrolles,S.Dejean,A.Legarra,H.Gilbert,
T.Druet,F.Ytournel,D.Estivals,N.Oumouhou and B.Mangin,
Combining linkage analysis and linkage disequilibrium for QTL fine mapping in animal pedigrees.}

\author{S. Dejean, N. Oumouhou, D. Estivals, B. Mangin }

\keyword{models}

