\name{HAPim-package}

\alias{HAPim-package}

\alias{HAPim}

\docType{package}

\title{HAPim}

\description{
The package provides a set of functions whose aim is to propose 4 methods of QTL detection:

-HAPimLD is an interval-mapping method designed for unrelated individuals 
with no family information that makes use of linkage disequilibrium.

-HAPimLDL is an interval-mapping method for design of half-sib families. It 
combines linkage analysis and linkage disequilibrium.

-HaploMax is  based on an analysis of variance with a dose haplotype effect. 

-HaploMaxHS is based on an analysis of variance with a sire effect and a dose haplotype effect in half-sib family design.

Fundings for the package development were provided to the LDLmapQTL project by the ANR GENANIMAL program and APIS-GENE.

}


\details{
\tabular{ll}{
Package: \tab HAPim\cr
Type: \tab Package\cr
Version: \tab 1.2\cr
Date: \tab 2008-08-25\cr
License: \tab none\cr
}
}

\author{S. Dejean, N. Oumouhou, D. Estivals, B. Mangin 

Maintainer: Brigitte Mangin <brigitte.mangin@toulouse.inra.fr> }

\references{publication to be submitted: C. Cierco-Ayrolles, S. Dejean, A. Legarra, H. Gilbert,
T. Druet, F. Ytournel, D. Estivals, N. Oumouhou and B. Mangin.
Combining linkage analysis and linkage disequilibrium for QTL fine mapping in animal pedigrees.}


\keyword{models}



